LARAVEL-TEST-APP-15 - Create Separate Controller and Service for Email Update - Created new EmailAddressController and Service - Updated user update routes
