+++ app/Http/Controllers/User/Profile/ProfileController.php
@@ -4,10 +4,7 @@ namespace App\Http\Controllers\User\Profile;
 
 use App\Enums\common\UserGuard;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Profile\{
-    UpdateEmailRequest,
-    UpdateProfileRequest,
-};
+use App\Http\Requests\Profile\UpdateProfileRequest;
 use App\Interfaces\{
     Job\JobInterface,
     User\UserInterface,
@@ -22,7 +19,6 @@ use Illuminate\Support\Facades\{
     Auth,
     DB,
 };
-use Symfony\Component\HttpFoundation\Response;
 
 class ProfileController extends Controller
 {
@@ -89,32 +85,6 @@ class ProfileController extends Controller
     }
 
     /**
-     * View user email address update page.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return \Illuminate\Contracts\View\View|\Illuminate\Http\RedirectResponse|void
-     *
-     * @throws \Exception
-     */
-    public function viewUpdateEmailPage()
-    {
-        try {
-            return view('user.update-email.index');
-        } catch (\Exception $e) {
-            LogService::error(
-                'Error showing update email address page.',
-                [
-                    'error' => $e->getMessage(),
-                    'trace' => $e->getTraceAsString(),
-                ],
-            );
-
-            throw $e;
-        }
-    }
-
-    /**
      * Update the user personal information.
      *
      * @param \App\Http\Requests\Profile\UpdateProfileRequest $request
@@ -156,84 +126,6 @@ class ProfileController extends Controller
             );
 
             return back()->with('error', $errorMessage);
-        }
-    }
-
-    /**
-     * Sends the verification link to the provided email address.
-     *
-     * @param \App\Http\Requests\Profile\UpdateProfileRequest $request
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function sendEmailAddressLink(UpdateEmailRequest $request): RedirectResponse
-    {
-        DB::beginTransaction();
-
-        try {
-            $newEmail = $request->validated()['email'];
-
-            $this->profileService->handleSendEmailAddressLink($newEmail);
-
-            DB::commit();
-
-            return back()->with('success', __('message.success.a_link_has_been_sent_to_the_provided_email'));
-        } catch (\Exception $e) {
-            DB::rollBack();
-
-            LogService::error(
-                'Error updating the user email address.',
-                [
-                    'error' => $e->getMessage(),
-                    'trace' => $e->getTraceAsString(),
-                ],
-            );
-
-            return back()->with('error', __('message.error.failed_handling_the_process'));
-        }
-    }
-
-    /**
-     * Update the user email address.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function updateEmailAddress(Request $request): RedirectResponse
-    {
-        $user = collect(Auth::guard(UserGuard::USER->value)->user())
-            ->only(['id', 'email']);
-
-        // Abort if the verification link is reused or invalid
-        abort_if($user['email'] === $request->new_email, Response::HTTP_FORBIDDEN);
-        abort_if(!$request->hasValidSignature(), Response::HTTP_FORBIDDEN);
-
-        DB::beginTransaction();
-
-        try {
-            $this->profileService->handleUpdateEmailAddress(
-                $request->new_email,
-                $user['id'],
-            );
-
-            DB::commit();
-
-            return redirect()
-                ->route('user.dashboard')
-                ->with('success', __('message.success.your_email_address_has_been_updated'));
-        } catch (\Exception $e) {
-            DB::rollBack();
-
-            LogService::error(
-                'Error updating the user email address.',
-                [
-                    'error' => $e->getMessage(),
-                    'trace' => $e->getTraceAsString(),
-                ],
-            );
-
-            return back()->with('error', __('message.error.failed_handling_the_process'));
         }
     }
 }
